+++ Libplanet/Store/BaseStore.cs
@@ -91,6 +91,23 @@ namespace Libplanet.Store
             IImmutableSet<Address> addressesToStrip)
             where T : IAction, new();
 
+        /// <inheritdoc/>
+        public abstract long GetTxNonce(string @namespace, Address address);
+
+        /// <inheritdoc/>
+        public abstract void IncreaseTxNonce<T>(
+            string @namespace,
+            Block<T> block)
+            where T : IAction, new();
+
+        /// <inheritdoc/>
+        public abstract void ForkTxNonce<T>(
+            string sourceNamespace,
+            string destinationNamespace,
+            Block<T> branchPoint,
+            IImmutableSet<Address> addressesToStrip)
+            where T : IAction, new();
+
         public long CountTransactions()
         {
             return IterateTransactionIds().LongCount();